Write functions for manage precommitData for 2-depth invoke
