+++ core/context.go
@@ -17,11 +17,11 @@ const (
 	NumDelegate         = 10
 	AccountDBNameFormat = "calculate_%d_%d_%d"
 	BackupDBNamePrefix  = "backup_"
-	BackupDBNameFormat  = BackupDBNamePrefix + "%d_%d"	// backup_CalcBH_accountDBIndex
+	BackupDBNameFormat  = BackupDBNamePrefix + "%d_%d" // backup_CalcBH_accountDBIndex
 
-	Revision8 uint64    = 8
-	RevisionMin = Revision8
-	RevisionMax = Revision8
+	Revision8   uint64 = 8
+	RevisionMin        = Revision8
+	RevisionMax        = Revision8
 )
 
 type IScoreDB struct {
@@ -32,12 +32,13 @@ type IScoreDB struct {
 	management    db.Database
 	calcResult    db.Database
 	preCommit     db.Database
+	preCommitInfo db.Database
 	claim         db.Database
 	claimBackup   db.Database
 
-	accountLock   sync.RWMutex
-	Account0      []db.Database
-	Account1      []db.Database
+	accountLock sync.RWMutex
+	Account0    []db.Database
+	Account1    []db.Database
 }
 
 func (idb *IScoreDB) getQueryDBList() []db.Database {
@@ -145,7 +146,7 @@ func (idb *IScoreDB) resetAccountDB(blockHeight uint64, oldCalcBH uint64) error 
 	}
 
 	// delete old backup account DB
-	oldBackup := filepath.Join(idb.info.DBRoot, BackupDBNamePrefix + strconv.FormatUint(oldCalcBH, 10) + "_*")
+	oldBackup := filepath.Join(idb.info.DBRoot, BackupDBNamePrefix+strconv.FormatUint(oldCalcBH, 10)+"_*")
 	oldBackups, err := filepath.Glob(oldBackup)
 	if err != nil {
 		log.Printf("Failed to get old backup account DB %s. %v", oldBackup, err)
@@ -168,7 +169,7 @@ func (idb *IScoreDB) resetAccountDB(blockHeight uint64, oldCalcBH uint64) error 
 		dbPath := filepath.Join(idb.info.DBRoot, dbName)
 		backupPath := filepath.Join(idb.info.DBRoot, fmt.Sprintf(BackupDBNameFormat, blockHeight, i+1))
 
-		_ , err := os.Stat(backupPath)
+		_, err := os.Stat(backupPath)
 		if os.IsNotExist(err) {
 			// backup old query DB
 			err = os.Rename(dbPath, backupPath)
@@ -182,7 +183,7 @@ func (idb *IScoreDB) resetAccountDB(blockHeight uint64, oldCalcBH uint64) error 
 		// open new calculate DB
 		newCalcDBs[i] = db.Open(idb.info.DBRoot, idb.info.DBType, dbName)
 	}
-	backup := filepath.Join(idb.info.DBRoot, BackupDBNamePrefix + strconv.FormatUint(blockHeight, 10) + "_*")
+	backup := filepath.Join(idb.info.DBRoot, BackupDBNamePrefix+strconv.FormatUint(blockHeight, 10)+"_*")
 	log.Printf("backup %d account DBs. %s", backupCount, backup)
 
 	// set new calculate DB
@@ -266,7 +267,7 @@ func (idb *IScoreDB) rollbackAccountDB(blockHeight uint64) error {
 		calcDBPostFix = 1
 	}
 
-	backups, err := filepath.Glob(filepath.Join(idb.info.DBRoot, BackupDBNamePrefix + "*"))
+	backups, err := filepath.Glob(filepath.Join(idb.info.DBRoot, BackupDBNamePrefix+"*"))
 	if err != nil {
 		log.Printf("Failed to get backup account DB")
 		return err
@@ -317,15 +318,16 @@ func (idb *IScoreDB) rollbackAccountDB(blockHeight uint64) error {
 }
 
 type Context struct {
-	DB              *IScoreDB
+	DB *IScoreDB
 
-	Revision        uint64
-	PRep            []*PRep
-	PRepCandidates  map[common.Address]*PRepCandidate
-	GV              []*GovernanceVariable
+	Revision       uint64
+	PRep           []*PRep
+	PRepCandidates map[common.Address]*PRepCandidate
+	GV             []*GovernanceVariable
 
-	stats    *Statistics
-	Rollback *Rollback
+	stats         *Statistics
+	Rollback      *Rollback
+	PreCommitInfo *PreCommitInfo
 }
 
 func (ctx *Context) getGVByBlockHeight(blockHeight uint64) *GovernanceVariable {
@@ -345,8 +347,8 @@ func (ctx *Context) UpdateGovernanceVariable(gvList []*IISSGovernanceVariable) {
 	// Update GV
 	for _, gvIISS := range gvList {
 		// there is new GV
-		if  len(ctx.GV) == 0 || ctx.GV[len(ctx.GV)-1].BlockHeight < gvIISS.BlockHeight {
-			gv :=  NewGVFromIISS(gvIISS)
+		if len(ctx.GV) == 0 || ctx.GV[len(ctx.GV)-1].BlockHeight < gvIISS.BlockHeight {
+			gv := NewGVFromIISS(gvIISS)
 
 			// write to memory
 			ctx.GV = append(ctx.GV, gv)
@@ -361,7 +363,7 @@ func (ctx *Context) UpdateGovernanceVariable(gvList []*IISSGovernanceVariable) {
 	gvLen := len(ctx.GV)
 	deleteOld := false
 	deleteIndex := -1
-	for i := gvLen - 1; i >= 0 ; i-- {
+	for i := gvLen - 1; i >= 0; i-- {
 		if ctx.GV[i].BlockHeight <= ctx.DB.getPrevCalcDoneBH() {
 			if deleteOld {
 				// delete from management DB
@@ -396,7 +398,7 @@ func (ctx *Context) UpdatePRep(prepList []*PRep) {
 	prepLen := len(ctx.PRep)
 	deleteOld := false
 	deleteIndex := -1
-	for i := prepLen - 1; i >= 0 ; i-- {
+	for i := prepLen - 1; i >= 0; i-- {
 		if ctx.PRep[i].BlockHeight <= ctx.DB.getPrevCalcDoneBH() {
 			if deleteOld {
 				// delete from management DB
@@ -483,7 +485,7 @@ func (ctx *Context) RollbackManagementDB(blockHeight uint64) {
 	// Rollback Governance Variable
 	bucket, _ := ctx.DB.management.GetBucket(db.PrefixGovernanceVariable)
 	gvLen := len(ctx.GV)
-	for i := gvLen - 1; i >= 0 ; i-- {
+	for i := gvLen - 1; i >= 0; i-- {
 		if ctx.GV[i].BlockHeight > blockHeight {
 			// delete from management DB
 			bucket.Delete(ctx.GV[i].ID())
@@ -495,7 +497,7 @@ func (ctx *Context) RollbackManagementDB(blockHeight uint64) {
 	// Rollback Main/Sub P-Rep list
 	bucket, _ = ctx.DB.management.GetBucket(db.PrefixPRep)
 	prepLen := len(ctx.PRep)
-	for i := prepLen - 1; i >= 0 ; i-- {
+	for i := prepLen - 1; i >= 0; i-- {
 		if ctx.PRep[i].BlockHeight > blockHeight {
 			// delete from management DB
 			bucket.Delete(ctx.PRep[i].ID())
@@ -561,7 +563,7 @@ func NewContext(dbPath string, dbType string, dbName string, dbCount int) (*Cont
 	}
 
 	// Open calculation result DB
-	isDB.calcResult= db.Open(isDB.info.DBRoot, isDB.info.DBType, "calculation_result")
+	isDB.calcResult = db.Open(isDB.info.DBRoot, isDB.info.DBType, "calculation_result")
 
 	// Open preCommit DB
 	isDB.preCommit = db.Open(isDB.info.DBRoot, isDB.info.DBType, "preCommit")
@@ -578,11 +580,14 @@ func NewContext(dbPath string, dbType string, dbName string, dbCount int) (*Cont
 	// make new Rollback stuff
 	ctx.Rollback = NewRollback()
 
+	// Open PreCommitHierarchy DB
+	isDB.preCommitInfo = db.Open(isDB.info.DBRoot, isDB.info.DBType, "preCommit_info")
+
 	return ctx, nil
 }
 
 func CloseIScoreDB(isDB *IScoreDB) {
-	log.Printf("Close 1 global DB and %d account DBs\n", len(isDB.Account0) + len(isDB.Account1))
+	log.Printf("Close 1 global DB and %d account DBs\n", len(isDB.Account0)+len(isDB.Account1))
 
 	// close management DB
 	isDB.management.Close()
